+++ app/requirements.txt
@@ -1,0 +1,19 @@
+calmap==0.0.11
+docling==2.42.1
+fastapi==0.116.1
+ipython==8.12.3
+langchain==0.3.26
+langchain_chroma==0.2.5
+langchain_core==0.3.71
+langchain_google_genai==2.1.8
+langchain_openai==0.3.28
+langchain_text_splitters==0.3.8
+langgraph==0.5.4
+matplotlib==3.10.3
+numpy==2.3.1
+pandas==2.3.1
+python-dotenv==1.1.1
+Requests==2.32.4
+seaborn==0.13.2
+typing_extensions==4.14.1
+uvicorn==0.35.0
