+++ docker-compose-vllm-embeddings.yml
@@ -16,7 +16,7 @@ services:
             - 8001:8001
         ipc: host
         image: vllm/vllm-openai:latest
-        command: --model /models/Qwen--Qwen3-Embedding-0.6B --port 8001 --max-num-seqs 8 --gpu-memory-utilization 0.8 --max_model_len=30000
+        command: --model /models/Qwen--Qwen3-Embedding-0.6B --port 8001 --max-num-seqs 8 --gpu-memory-utilization 0.8 --max_model_len=3000
     chatbot:
         build:
             context: .
@@ -28,6 +28,7 @@ services:
         volumes:
         - ./.env:/app/.env
         - ./app/app.py:/app/app.py
+        - ./app/router_logic.py:/app/router_logic.py
         - ./app/logic/query_logic.py:/app/query_logic.py
         - ./app/RAG/get_embedding_function.py:/app/get_embedding_function.py
         - ./app/RAG/populate_database.py:/app/populate_database.py
