+++ docker-compose-vllm.yml
@@ -58,6 +58,7 @@ services:
         volumes:
         - ./.env:/app/.env
         - ./app/app.py:/app/app.py
+        - ./app/router_logic.py:/app/router_logic.py
         - ./app/logic/query_logic.py:/app/query_logic.py
         - ./app/RAG/get_embedding_function.py:/app/get_embedding_function.py
         - ./app/RAG/populate_database.py:/app/populate_database.py
