+++ logs/chat_logs.csv
@@ -29,3 +29,4 @@ nuevo@ejemplo.com,que es un algoritmo genético?,09:58:15,2025-07-21,metaheurist
 nuevo@ejemplo.com,que es un algoritmo genético?,10:01:49,2025-07-21,metaheuristicas,"Sem01-Problemas-MHs.pdf:37:215,Tema01-Metaheuristicas-24-25.pdf:6:142,Tema02-Modelos de Búsqueda-Trayectorias-vs-Poblaciones-24-25.pdf:22:63,Sem01-Problemas-MHs.pdf:35:213,Sem01-Problemas-MHs.pdf:50:228"
 nuevo@ejemplo.com,que es un algoritmo genético?,12:24:45,2025-07-21,metaheuristicas,"Sem01-Problemas-MHs.pdf:11:195,Sem01-Problemas-MHs.pdf:55:233,Tema02-Modelos de Búsqueda-Trayectorias-vs-Poblaciones-24-25.pdf:13:55,Guion_P1a_LocalGreedy_MDD_MHs_2024-25.pdf:4:133,Tema02-Modelos de Búsqueda-Trayectorias-vs-Poblaciones-24-25.pdf:12:54"
 nuevo@ejemplo.com,que es un algoritmo genético?,12:37:41,2025-07-21,metaheuristicas,"Sem01-Problemas-MHs.pdf:11:195,Sem01-Problemas-MHs.pdf:11:164,Sem01-Problemas-MHs.pdf:55:233,Sem01-Problemas-MHs.pdf:55:201,Tema02-Modelos de Búsqueda-Trayectorias-vs-Poblaciones-24-25.pdf:13:55"
+nuevo@ejemplo.com,que es MDD?,09:30:22,2025-07-22,metaheuristicas,"Guion_P1a_LocalGreedy_MDD_MHs_2024-25.pdf:0:125,Guion_P1a_LocalGreedy_MDD_MHs_2024-25.pdf:0:100,Tema02-Modelos de Búsqueda-Trayectorias-vs-Poblaciones-24-25.pdf:35:76,Tema02-Modelos de Búsqueda-Trayectorias-vs-Poblaciones-24-25.pdf:35:63,Guion_P1a_LocalGreedy_MDD_MHs_2024-25.pdf:4:133"
