+++ requirements.txt
@@ -1,32 +1,39 @@
-accelerate==1.8.1
-aiohappyeyeballs==2.6.1
-aiohttp==3.12.14
-aiosignal==1.4.0
+accelerate==1.9.0
 annotated-types==0.7.0
 anyio==4.9.0
+asttokens==3.0.0
 attrs==25.3.0
+backcall==0.2.0
 backoff==2.2.1
 bcrypt==4.3.0
-bitsandbytes==0.46.1
+beautifulsoup4==4.13.4
 build==1.2.2.post1
 cachetools==5.5.2
-certifi==2025.7.9
-cffi==1.17.1
+calmap==0.0.11
+certifi==2025.7.14
 charset-normalizer==3.4.2
 chromadb==1.0.15
 click==8.2.1
 coloredlogs==15.0.1
-cryptography==45.0.5
-datasets==4.0.0
-dill==0.3.8
+contourpy==1.3.2
+cycler==0.12.1
+decorator==5.2.1
+dill==0.4.0
 distro==1.9.0
+docling==2.42.1
+docling-core==2.43.0
+docling-ibm-models==3.8.2
+docling-parse==4.1.0
 durationpy==0.10
+easyocr==1.7.2
+et-xmlfile==2.0.0
+executing==2.2.0
 fastapi==0.116.1
 filelock==3.18.0
 filetype==1.2.0
 flatbuffers==25.2.10
-frozenlist==1.7.0
-fsspec==2025.3.0
+fonttools==4.59.0
+fsspec==2025.7.0
 google-ai-generativelanguage==0.6.18
 google-api-core==2.25.1
 google-auth==2.40.3
@@ -42,35 +49,48 @@ httpx==0.28.1
 huggingface-hub==0.33.4
 humanfriendly==10.0
 idna==3.10
+imageio==2.37.0
 importlib-metadata==8.7.0
 importlib-resources==6.5.2
+ipython==8.12.3
+jedi==0.19.2
 jinja2==3.1.6
 jiter==0.10.0
+jsonlines==3.1.0
 jsonpatch==1.33
 jsonpointer==3.0.0
-jsonschema==4.24.0
+jsonref==1.1.0
+jsonschema==4.25.0
 jsonschema-specifications==2025.4.1
+kiwisolver==1.4.8
 kubernetes==33.1.0
 langchain==0.3.26
-langchain-chroma==0.2.4
-langchain-core==0.3.68
+langchain-chroma==0.2.5
+langchain-core==0.3.71
 langchain-google-genai==2.1.8
-langchain-openai==0.3.27
+langchain-openai==0.3.28
 langchain-text-splitters==0.3.8
-langgraph==0.5.3
-langgraph-checkpoint==2.1.0
+langgraph==0.5.4
+langgraph-checkpoint==2.1.1
 langgraph-prebuilt==0.5.2
-langgraph-sdk==0.1.73
-langsmith==0.4.5
+langgraph-sdk==0.1.74
+langsmith==0.4.8
+latex2mathml==3.78.0
+lazy-loader==0.4
+lxml==5.4.0
 markdown-it-py==3.0.0
+marko==2.1.4
 markupsafe==3.0.2
+matplotlib==3.10.3
+matplotlib-inline==0.1.7
 mdurl==0.1.2
 mmh3==5.1.0
+mpire==2.10.2
 mpmath==1.3.0
-multidict==6.6.3
-multiprocess==0.70.16
-networkx==3.5
-numpy==2.3.1
+multiprocess==0.70.18
+networkx==3.4.2
+ninja==1.11.1.4
+numpy==2.2.6
 nvidia-cublas-cu12==12.6.4.1
 nvidia-cuda-cupti-cu12==12.6.80
 nvidia-cuda-nvrtc-cu12==12.6.77
@@ -87,41 +107,51 @@ nvidia-nvjitlink-cu12==12.6.85
 nvidia-nvtx-cu12==12.6.77
 oauthlib==3.3.1
 onnxruntime==1.22.1
-openai==1.95.0
+openai==1.97.1
+opencv-python-headless==4.11.0.86
+openpyxl==3.1.5
 opentelemetry-api==1.35.0
 opentelemetry-exporter-otlp-proto-common==1.35.0
 opentelemetry-exporter-otlp-proto-grpc==1.35.0
 opentelemetry-proto==1.35.0
 opentelemetry-sdk==1.35.0
 opentelemetry-semantic-conventions==0.56b0
-orjson==3.10.18
+orjson==3.11.0
 ormsgpack==1.10.0
 overrides==7.7.0
-packaging==24.2
+packaging==25.0
 pandas==2.3.1
-pdfminer-six==20250506
-pdfplumber==0.11.7
-peft==0.16.0
+parso==0.8.4
+pexpect==4.9.0
+pickleshare==0.7.5
 pillow==11.3.0
+pluggy==1.6.0
 posthog==5.4.0
-propcache==0.3.2
+prompt-toolkit==3.0.51
 proto-plus==1.26.1
 protobuf==6.31.1
 psutil==7.0.0
-pyarrow==20.0.0
+ptyprocess==0.7.0
+pure-eval==0.2.3
 pyasn1==0.6.1
 pyasn1-modules==0.4.2
 pybase64==1.4.1
-pycparser==2.22
+pyclipper==1.3.0.post6
 pydantic==2.11.7
 pydantic-core==2.33.2
+pydantic-settings==2.10.1
 pygments==2.19.2
-pypdfium2==4.30.1
+pylatexenc==2.10
+pyparsing==3.2.3
+pypdfium2==4.30.0
 pypika==0.48.9
 pyproject-hooks==1.2.0
+python-bidi==0.6.6
 python-dateutil==2.9.0.post0
+python-docx==1.2.0
 python-dotenv==1.1.1
 python-multipart==0.0.20
+python-pptx==1.0.2
 pytz==2025.2
 pyyaml==6.0.2
 referencing==0.36.2
@@ -132,20 +162,30 @@ requests-toolbelt==1.0.0
 rich==14.0.0
 rpds-py==0.26.0
 rsa==4.9.1
+rtree==1.4.0
 safetensors==0.5.3
+scikit-image==0.25.2
+scipy==1.15.3
+seaborn==0.13.2
+semchunk==2.2.2
 setuptools==80.9.0
+shapely==2.1.1
 shellingham==1.5.4
 six==1.17.0
 sniffio==1.3.1
+soupsieve==2.7
 sqlalchemy==2.0.41
-starlette==0.47.1
+stack-data==0.6.3
+starlette==0.47.2
 sympy==1.14.0
+tabulate==0.9.0
 tenacity==9.1.2
+tifffile==2025.6.11
 tiktoken==0.9.0
 tokenizers==0.21.2
-torch==2.7.1
 tqdm==4.67.1
-transformers==4.53.2
+traitlets==5.14.3
+transformers==4.53.3
 triton==3.3.1
 typer==0.16.0
 typing-extensions==4.14.1
@@ -155,9 +195,10 @@ urllib3==2.5.0
 uvicorn==0.35.0
 uvloop==0.21.0
 watchfiles==1.1.0
+wcwidth==0.2.13
 websocket-client==1.8.0
 websockets==15.0.1
+xlsxwriter==3.2.5
 xxhash==3.5.0
-yarl==1.20.1
 zipp==3.23.0
 zstandard==0.23.0
