+++ tests/parsed_outputDoclings/modelos_avanzados_computacion/EjemploReduccion.txt
@@ -1,0 +1,114 @@
+/BX/D0
+
+## /C8/D6/D3/CQ/D0/CT/D1/CP
+
+/BV/B9/CA/BX/BZ/CD/C4/BT/CA/B4/C5/B5
+
+/BX/D0 /C8/D6/D3/CQ/D0/CT/D1/CP /BV/B9/CA/BX/BZ/CD/C4/BT/CA/B4/C5/B5 /BW/CP/CS/CP /D9/D2/CP /C5/CC M /B8 /CS/CT/D8/CT/D6/D1/CX/D2/CP/D6 /D7/CX L ( M ) /D2/D3 /CT/D7 /D6/CT/CV/D9/D0/CP/D6 /B4/CT/D0
+
+/D3/D1/D4/D0/CT/D1/CT/D2/D8/CP/D6/CX/D3
+
+/CS/CT/D0
+
+/D4/D6/D3/CQ/D0/CT/D1/CP
+
+/CA/BX/BZ/CD/C4/BT/CA/BA
+
+/BX/D0
+
+## /C8/D6/D3/CQ/D0/CT/D1/CP
+
+/BV/B9/CA/BX/BZ/CD/C4/BT/CA/B4/C5/B5
+
+/BX/D0 /C8/D6/D3/CQ/D0/CT/D1/CP /BV/B9/CA/BX/BZ/CD/C4/BT/CA/B4/C5/B5 /BW/CP/CS/CP /D9/D2/CP /C5/CC M /B8 /CS/CT/D8/CT/D6/D1/CX/D2/CP/D6 /D7/CX L ( M ) /D2/D3 /CT/D7 /D6/CT/CV/D9/D0/CP/D6 /B4/CT/D0
+
+/C6/D3 /CT/D7 /CB/CT/D1/CX/CS/CT/CX/CS/CX/CQ/D0/CT /CB/CT /D4/D9/CT/CS/CT /CW/CP/CT/D6 /D6/CT/CS/D9/CX/CT/D2/CS/D3 /CT/D0 /D4/D6/D3/CQ/D0/CT/D1/CP /BV/B9/CD/C6/C1/CE/BX/CA/CB/BT/C4/B4/C5/B8/D9/B5 /B4/CT/D0
+
+## /D3/D1/D4/D0/CT/D1/CT/D2/D8/CP/D6/CX/D3 /CS/CT/D0
+
+/D4/D6/D3/CQ/D0/CT/D1/CP
+
+/CA/BX/BZ/CD/C4/BT/CA/BA
+
+/D3/D1/D4/D0/CT/D1/CT/D2/D8/CP/D6/CX/D3
+
+/CS/CT/D0
+
+/D4/D6/D3/CQ/D0/CT/D1/CP
+
+/D9/D2/CX/DA/CT/D6/D7/CP/D0/B5
+
+/CP
+
+/BV/B9/CA/BX/BZ/CD/C4/BT/CA/B4/C5/B3/B5
+
+/BX/D0
+
+## /C8/D6/D3/CQ/D0/CT/D1/CP
+
+/BV/B9/CA/BX/BZ/CD/C4/BT/CA/B4/C5/B5
+
+/BX/D0 /C8/D6/D3/CQ/D0/CT/D1/CP /BV/B9/CA/BX/BZ/CD/C4/BT/CA/B4/C5/B5 /BW/CP/CS/CP /D9/D2/CP /C5/CC M /B8 /CS/CT/D8/CT/D6/D1/CX/D2/CP/D6 /D7/CX L ( M ) /D2/D3 /CT/D7 /D6/CT/CV/D9/D0/CP/D6 /B4/CT/D0
+
+/C6/D3 /CT/D7 /CB/CT/D1/CX/CS/CT/CX/CS/CX/CQ/D0/CT /CB/CT /D4/D9/CT/CS/CT /CW/CP/CT/D6 /D6/CT/CS/D9/CX/CT/D2/CS/D3 /CT/D0 /D4/D6/D3/CQ/D0/CT/D1/CP /BV/B9/CD/C6/C1/CE/BX/CA/CB/BT/C4/B4/C5/B8/D9/B5 /B4/CT/D0
+
+## /D3/D1/D4/D0/CT/D1/CT/D2/D8/CP/D6/CX/D3 /CS/CT/D0
+
+/D4/D6/D3/CQ/D0/CT/D1/CP
+
+/CA/BX/BZ/CD/C4/BT/CA/BA
+
+## /D3/D1/D4/D0/CT/D1/CT/D2/D8/CP/D6/CX/D3 /CS/CT/D0
+
+/D4/D6/D3/CQ/D0/CT/D1/CP
+
+/D9/D2/CX/DA/CT/D6/D7/CP/D0/B5
+
+/CP
+
+/BV/B9/CA/BX/BZ/CD/C4/BT/CA/B4/C5/B3/B5
+
+/CA/CT/CS/D9/CX/D6
+
+/CA/CT/CS/D9/CX/GR/D2 /BX/D5/D9/CX/DA/CP/D0/CT/D2/D8/CT
+
+/BV/B9/CD/C6/C1/CE/BX/CA/CB/BT/C4/B4/C5/B8/D9/B5
+
+/CP
+
+/BV/B9/CA/BX/BZ/CD/C4/BT/CA/B4/C5/B3/B5
+
+/CT/D7
+
+/CT/D5/D9/CX/DA/CP/D0/CT/D2/D8/CT
+
+/CA/CT/CS/D9/CX/D6
+
+/CD/C6/C1/CE/BX/CA/CB/BT/C4/B4/C5/B8/D9/B5
+
+/CP
+
+/CA/BX/BZ/CD/C4/BT/CA/B4/C5/B3/B5
+
+/CP
+
+/CA/CT/CS/D9/CX/GR/D2 /CS/CT /CD/C6/C1/CE/BX/CA/CB/BT/C4/B4/C5/B8/D9/B5 /CP /CA/BX/BZ/CD/C4/BT/CA/B4/C5/B3/B5 /BW/CP/CS/D3/D7 /D9/D2/D3/D7 /CS/CP/D8/D3/D7 /CS/CT/D0 /CD/C6/C1/CE/BX/CA/CB/BT/C4/B8 ( M , u ) /B8 /D3/D2/D7/D8/D6/D9/GL/D1/D3/D7 /D9/D2/D3/D7 /CS/CP/D8/D3/D7 M ′
+
+/CS/CT
+
+/CA/BX/BZ/CD/C4/BT/CA/B8
+
+/CS/CT
+
+/D0/CP
+
+/D7/CX/CV/D9/CX/CT/D2/D8/CT
+
+/CU/D3/D6/D1/CP/BM
+
+<!-- image -->
+
+M , { n n : m ≥ }
+
+- /CB/CX /CP/CT/D4/D8/CP /B4/CB/C1/B5/B8 /D7/CT/CP /CT/D0 /D2/GY/D1/CT/D6/D3 /CS/CT /D4/CP/D7/D3/D7 /D5/D9/CT /CS/CP /CT/D2 /CT/D7/D8/CT /CP/CT/D4/D8/CP/CX/GR/D2/BA /BX/D2/D8/D3/D2/CT/D7/B8 /BC /BD /BE /B8 /D5/D9/CT /D3/D1/D3 /CT/D7 /AS/D2/CX/D8/D3 /CT/D7 /D6/CT/CV/D9/D0/CP/D6 /B4/CB/C1/B5 /CB/CX M /D2/D3 /CP/CT/D4/D8/CP u /B4/C6/C7/B5/B8 /CT/D2/D8/D3/D2/CT/D7 L ( M ) = { /BC n /BD n : n ≥ /BC } /B8 /D5/D9/CT /D7/CP/CQ/CT/D1/D3/D7 /D5/D9/CT /D2/D3
+- /BW/D3/D2/CS/CT /BC /BD /CT/D7 /D9/D2/CP /C5/CC /D5/D9/CT /CP/CT/D4/D8/CP /BC /BD /BC /BA M u n L ( M ′ ) = { k k : k &lt; n }
