+++ tests/parsed_outputUnstructured/metaheuristicas/Sem02-ProblemaSNIM-BusquedaLocal-MH.txt
@@ -1,0 +1,1 @@
+1. Problema de Maximizar la Influencia en Redes Sociales (SNIMP) Social Network Influence Maximization Problem, Márketing digital. Modelos epidémicos El Problema de Maximizar la Influencia en Redes Sociales, Social Networks Influence Maximization Problem, SNIMP, es un problema de optimización combinatoria con una formulación sencilla pero una resolución compleja (es NP-completo). El problema general consiste en seleccionar un subconjunto Sel de m inicial S de n elementos elementos (Mm) de un conjunto (obviamente, n m) de forma que se maximice la influencia entre los elementos escogidos. Además de los n elementos (ei, i1,...,n) y el número de elementos a seleccionar m, se dispone de una matriz C(Cij) de dimensión nn no simétrica que contiene las conexiones entre los nodos de la red. C puede ser bastante dispersa, con un número muy variable de Para el problema con el que trabajaremos en prácticas, se busca lo siguiente: ssSmax ICM(G,S, p,ev) G es el grafo del problema. S es el conjunto de posibles combinaciones. p es la probabilidad de influencia/contagio (p0.01) ev es el número de simulación del ICM (ev10) A es el conjunto de nodos activados, I su tamaño I. Lozano-Osorio, J. Sánchez-Oro, A. Duarte, y O. Cordón, A quick GRASP-based method for influence maximization in social networks, J Ambient Intell Human Comput, vol. 14, n. 4, pp. 3767- 3779, abr. 2023, doi: 10.1007/s12652-021-03510-4. La mayoría de propuestas abordadas son Greedy. Hay varias definiciones. Vamos a usar una de las más sencillas que ofrece un buen resultado según experimentos: ALGORITMO GREEDY: 1) Se calcula para cada nodo u el número de nodos vecinos: d(u)Nu ,Nu wV :(u,v)E 2) Se define como valor heurístico de un nodo el total de vecinos, y vecinos de éstos: g(u)du s vNu du 3) Se listan los valores posibles ordenados por su valor heurístico. 4) Se escogen los k primeros. Búsquedas por Trayectorias Simples: Búsqueda Local del Mejor Representación: Problema de selección: un conjunto Sel{s1, , sm} que almacena los m elementos seleccionados de entre los n elementos del conjunto S Para ser una solución candidata válida, tiene que satisfacer las restricciones (ser un conjunto de tamaño m): No puede tener elementos repetidos Ha de contener exactamente m elementos El orden de los elementos no es relevante Búsquedas por Trayectorias Simples: Búsqueda Local del Mejor Operador de vecino de intercambio y su entorno: El entorno de una solución Sel está formado por las soluciones accesibles desde ella a través de un movimiento de intercambio Dada una solución (conjunto de elementos seleccionados) se escoge un elemento y se intercambia por otro que no estuviera seleccionado (Int(Sel,i,j)): Sel {s1, , i, , sm} Sel {s1, , j, , sm} Int(Sel,i,j) verifica las restricciones: si la solución original Sel es factible y el elemento j se escoge de los no seleccionados en Sel, es decir, del conjunto S-Sel, siempre genera una solución vecina Sel factible Búsquedas por Trayectorias Simples: Búsqueda Local del Mejor Su aplicación provoca que el tamaño del entorno sea demasiado grande (m!), m10 más de 3 millones combinaciones. La BL del Mejor del MDP explora todo el vecindario, las soluciones resultantes de los m(n-m) intercambios posibles, escoge el mejor vecino y se mueve a él siempre que se produzca mejora Si no la hay, detiene la ejecución y devuelve la solución actual El método funciona bien pero es muy lento incluso para casos no Vamos a probar también parar cuando tras Nvecinos no mejore (Nvecinos50). Se utilizarán 4 casos reales seleccionados de varios de los conjuntos de instancias del Stanford Large Network Dataset Collection:  Se podría usar esos ficheros, pero hemos modificado y puesto en PRADO una versión sin los nodos sin conexiones, para agilizar y reducir el uso de memoria. No planteamos el óptimo, porque en la mayoría de los casos no se conocen, Para el MDDP, disponemos de cuatro conjunto de datos: Ca-GrCQc: Relación de autores en revista científica, tiene 5 242 nodos y 14 P2p-Gnutella05: Intercambio de mensajes en una red social el 5 Agosto 2002, tiene P2p-Gnutella08: Intercambio de mensajes en una red social el 8 Agosto 2002, tiene P2p-Gnutella25: Intercambio de mensajes en una red social, 25 Agosto 2002, tiene El formato de los ficheros de datos es un fichero de texto con la siguiente cabecera: Directed graph (each unordered pair of nodes is saved once): NombreFichero.txt Descripción Nodes: N Edges: E FromNodeId ToNodeId Donde N es el número de elementos, E es el número de conexiones. A continuación aparecen los pares de valores (u, v) pertenecientes a E. La EJEMPLO: FICHERO DEL CASO ca-GrQc.txt: . FromNodeId ToNodeId 0 1 0 2 0 3 0 4 0 5 0 6 0 7 0 8 4 9 4 0 4 10 ... 4410 370 4974 4973 4974 4976 1189 1061 1189 1176 1189 1177 Para la preparación de las transparencias de presentación del Isaac Lozano-Osorio. Universidad Rey Juan Carlos Jesús Sánchez-Oro. Universidad Rey Juan Carlos Abraham Duarte. Universidad Rey Juan Carlos Óscar Cordón. Universidad de Granada. Otra referencia que me ha servido muy útil: Y. Ye, Y. Chen, y W. Han, Influence maximization in social networks: Theories, methods and challenges, Array, vol. 16, p. 100264, dic. 2022, doi: 10.1016/j.array.2022.100264. Además, agradecimientos a Óscar por introducirme al problema.
