+++ tests/parsed_outputUnstructured/modelos_avanzados_computacion/EjemploReduccion.txt
@@ -1,0 +1,1 @@
+El Problema C-REGULAR(M) El Problema C-REGULAR(M) Dada una MT M , determinar si L(M) no es regular (el El Problema C-REGULAR(M) El Problema C-REGULAR(M) Dada una MT M , determinar si L(M) no es regular (el Se puede ha er redu iendo el problema C-UNIVERSAL(M,u) (el omplementario del problema universal) a C-REGULAR(M') El Problema C-REGULAR(M) El Problema C-REGULAR(M) Dada una MT M , determinar si L(M) no es regular (el Se puede ha er redu iendo el problema C-UNIVERSAL(M,u) (el omplementario del problema universal) a C-REGULAR(M') Redu ir C-UNIVERSAL(M,u) a C-REGULAR(M') es equivalente a Redu ir UNIVERSAL(M,u) a REGULAR(M') Dados unos datos del UNIVERSAL, (M,u), onstruímos unos datos de REGULAR, M M w pasos es una MT que a epta { n : m 0}. Si M a epta u (SI), sea n el número de pasos que da en este a epta ión. k : 2k n}, que omo es (cid:28)nito es regular (SI) Enton es, L(M) {0 n Si M no a epta u (NO), enton es L(M) { n : n 0}, que sabemos que no
